task: patch codeql bug Implicit narrowing conversion in compound assignment

+++ engine/src/main/java/io/orqueio/bpm/engine/impl/jobexecutor/BackoffJobAcquisitionStrategy.java
@@ -292,7 +292,7 @@ public class BackoffJobAcquisitionStrategy implements JobAcquisitionStrategy {
     if (applyJitter) {
       // add a bounded random jitter to avoid multiple job acquisitions getting exactly the same
       // polling interval
-      backoffTime += Math.random() * (backoffTime / 2);
+      backoffTime += (long) (Math.random() * (backoffTime / 2));
     }
 
     return backoffTime;
